Update persistence configurations with @Import annotation
