+++ GiftCertificatesSystem/persistence/src/main/java/com/epam/esm/context/PersistenceConfig.java
@@ -1,0 +1,41 @@
+package com.epam.esm.context;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+
+import javax.sql.DataSource;
+
+@Configuration
+@PropertySource("classpath:datasource.properties")
+@ComponentScan("com.epam.esm.dao")
+public class PersistenceConfig {
+    @Autowired
+    private Environment environment;
+
+    @Bean
+    public PersistenceExceptionTranslationPostProcessor persistenceExceptionTranslationPostProcessor() {
+        return new PersistenceExceptionTranslationPostProcessor();
+    }
+
+    @Bean
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName(environment.getProperty("dataSource.driverClassName"));
+        dataSource.setUrl(environment.getProperty("dataSource.url"));
+        dataSource.setUsername(environment.getProperty("dataSource.username"));
+        dataSource.setPassword(environment.getProperty("dataSource.password"));
+        return dataSource;
+    }
+
+    @Bean
+    public JdbcTemplate jdbcTemplate() {
+        return new JdbcTemplate(dataSource());
+    }
+}
