+++ GiftCertificatesSystem/persistence/src/test/java/com/epam/esm/context/TestConfig.java
@@ -1,36 +1,22 @@
 package com.epam.esm.context;
 
-import com.epam.esm.dao.GiftCertificateDao;
-import com.epam.esm.dao.TagDao;
-import com.epam.esm.dao.impl.GiftCertificateDaoImpl;
-import com.epam.esm.dao.impl.TagDaoImpl;
 import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
 
 import javax.sql.DataSource;
 
+@Configuration
+@Import(PersistenceConfig.class)
 public class TestConfig {
     @Bean
+    @Profile("test")
     public DataSource dataSource() {
         return new EmbeddedDatabaseBuilder()
                 .setType(EmbeddedDatabaseType.H2)
                 .build();
-    }
-
-    @Bean
-    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-        return new JdbcTemplate(dataSource);
-    }
-
-    @Bean
-    public TagDao tagDao(JdbcTemplate jdbcTemplate) {
-        return new TagDaoImpl(jdbcTemplate);
-    }
-
-    @Bean
-    public GiftCertificateDao giftCertificateDao(JdbcTemplate jdbcTemplate) {
-        return new GiftCertificateDaoImpl(jdbcTemplate);
     }
 }
