+++ GiftCertificatesSystem/persistence/src/test/java/com/epam/esm/dao/impl/GiftCertificateDaoImplTest.java
@@ -22,7 +22,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = {TestConfig.class})
+@ContextConfiguration(classes = TestConfig.class)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @ActiveProfiles("test")
 @Sql(scripts = {"classpath:create_gift_certificate_table.sql"})
