+++ GiftCertificatesSystem/web/src/main/java/com/epam/esm/context/ApplicationConfig.java
@@ -1,53 +1,7 @@
 package com.epam.esm.context;
 
-import com.epam.esm.dao.GiftCertificateDao;
-import com.epam.esm.dao.TagDao;
-import com.epam.esm.dao.impl.GiftCertificateDaoImpl;
-import com.epam.esm.dao.impl.TagDaoImpl;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
-import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-
-import javax.sql.DataSource;
 
 @Configuration
-@PropertySource("classpath:datasource.properties")
 public class ApplicationConfig {
-    @Autowired
-    private Environment environment;
-
-    @Bean
-    public PersistenceExceptionTranslationPostProcessor persistenceExceptionTranslationPostProcessor() {
-        return new PersistenceExceptionTranslationPostProcessor();
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(environment.getProperty("dataSource.driverClassName"));
-        dataSource.setUrl(environment.getProperty("dataSource.url"));
-        dataSource.setUsername(environment.getProperty("dataSource.username"));
-        dataSource.setPassword(environment.getProperty("dataSource.password"));
-        return dataSource;
-    }
-
-    @Bean
-    public JdbcTemplate jdbcTemplate() {
-        return new JdbcTemplate(dataSource());
-    }
-
-    @Bean
-    public GiftCertificateDao giftCertificateDao() {
-        return new GiftCertificateDaoImpl(jdbcTemplate());
-    }
-
-    @Bean
-    public TagDao tagDao() {
-        return new TagDaoImpl(jdbcTemplate());
-    }
 }
